Pandas-Data-Science-rainfall in Tunisia
Setup

To access all of the files I recommend you to clone it locally (fork-a-repo) . Instructions on how to do this can be found here: https://help.github.com/en/github/getting-started-with-github/fork-a-repo

The other option is to click the green "clone or download" button and then click "Download ZIP". You then should extract all of the files to the location you want to edit your code.

Installing Jupyter Notebook: https://jupyter.readthedocs.io/en/latest/install.html
Installing Pandas library: https://pandas.pydata.org/pandas-docs/stable/install.html

Background Information:
In this project i use Python Pandas & Python Matplotlib to analyze rainfall in Tunisia in real time. The data is extracted from the server of Ministry of Agriculture 

I explore 3 questions related to the data:
what is  Highest Volume of rainfall in tunisia? in which station and which day?
what is in each year the record of rainfall in tunisia ?
which stations know the most rainfall level ?